+++ src/paths.ts
@@ -1,7 +1,15 @@
 import path from "path"
 import fs from "fs-extra"
 
-export const HOME = path.join(process.env.HOME as string, "origeen")
+// https://stackoverflow.com/a/26227660/13100860
+export const HOME = path.join(
+    process.env.APPDATA ||
+        (process.platform === "darwin"
+            ? `${process.env.HOME}/Library/Preferences`
+            : process.env.HOME + "/.local/share"),
+    "origeen"
+)
+
 export const CONFIG = path.join(HOME, "config.json")
 export const PROJECTS = path.join(HOME, "projects.json")
 export const TEMPLATES = path.join(HOME, "templates")
@@ -11,6 +19,10 @@ export const EMPTY_TEMPLATE = path.join(TEMPLATES, "@empty")
  * Ensures that all the exported constants points to an existing and valid file/directory
  */
 export function ensure() {
+    console.debug("Origeen's home:")
+    console.debug(HOME)
+    console.debug()
+
     fs.ensureDirSync(HOME)
     if (!fs.existsSync(CONFIG)) fs.writeJsonSync(CONFIG, {})
     if (!fs.existsSync(PROJECTS)) fs.writeJsonSync(PROJECTS, [])